tests/wpt/web-platform-tests/referrer-policy/generic/multiple-headers-one-invalid.html

<!DOCTYPE html>
<html>
  <head>
    <title>Referrer Policy: multiple Referrer-Policy headers with one invalid</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <script src="/common/security-features/resources/common.js"></script>
    <script src="/referrer-policy/generic/referrer-policy-test-case.js"></script>
  </head>
  <body>
    <h1>Referrer Policy: multiple Referrer-Policy headers with one invalid</h1>
    <p></p>

    <pre id="received_message">Running...</pre>

    <script>
      promise_test(() => {
        var urlPath = '/common/security-features/subresource/image.py?cache_destroyer=' + (new Date()).getTime();
        var url = location.protocol + "//www1." + location.hostname + ":" + location.port +
                  urlPath;
        return requestViaImageForReferrerPolicy(url, null, 'no-referrer')
          .then(function(message) {
              assert_equals(message.referrer, document.location.href);
            });
      }, "Referrer policy header parsing fails if one header is invalid");
    </script>

    <div id="log"></div>
  </body>
</html>
